Use the test_connect machinery to select the first address that we can connect to.
When resolving a name, you can often get multiple IP addresses to connect to, both ipv4 and
ipv6. Go through all of them, and pick the first one we are able to connect to as our
benchmarking target.

An alternative could be to use ALL of the IP addresses we find and go round robin against
them.

Renamedt test_connection to select_address and modified slightly to try all addresses
and return the first one that works.